WIP check-syntax: Improve racket--do-describe
Instead of the "Visit Definition" and "Documentation in Browser"
buttons being hard-coded assuming the REPL namespace, now they are
added using optional thunk arguments.

Have racket-check-syntax-describe supply suitable thunks, which are
effectively "lazy" variations of racket-check-syntax-visit-definition
and racket-check-syntax-documentation: Go ahead and check the
racket-mode buffer for properties, which are closed over in the thunks
and will be correct if/when the thunks are called from the user
clicking the buttons. After all, the racket-mode buffer won't be
selected when the *Racket Describe* is, and the latter might hang
around even after the user edits the former -- but it should still
work.