WIP check-syntax: Various changes for describe
Make describe able to use find-signature with the path-str flavor --
where the syntax is in our cache from using string->expanded-syntax.
This entails keeping non-expanded syntax in the cache, too. As well as
sorting out the various permutations of where/how to get the plain or
expanded syntax (from actual files or from the cache).

Also add/use a how/c contract.