Change "doc" command to return the URL, not open the browser
Note: Emacs' default browse-url might not make the browser open the
page positioned at the anchor a.k.a. fragment, on macOS. Previously we
had worked around this on macOS, down the Racket back-end. However
this would now need to happen up in Emacs. This commit doesn't supply
any such solution. I feel like it's out of scope for Racket Mode. If
anything, it should be a little package that fixes browse-url
generally for Emacs.

In any case, git history has the code we deleted, to (a) determine the
default browser and then (b) do

  osascript -e 'tell application "@browser" to open location "@file-url" activate'

The only non-trivial part is determining the default browser. Assuming
the user must supply that browser name, then the rest of it would be a
pretty simple function to write and set as the value of the variable
`browse-url-browser-function'.