Replace don't augment `racket-complete-at-point'
If we leave it active, it might return its "dumb" candidates that come
from font-lock keywords, even though they are not available based on
actual imports.

So remove the racket-complete-at-point hook when our minor mode
enables, and set it back again when we disable.

When adding either racket-complete-at-point or racket-xp-at-point, use
the true APPEND argument to `add-hook'. I think that's correct and
best? That means functions added by the user will get first crack?
Also it's possible we might implement #420 using a distinct completion
at point function, and, probably that should "go first" (if point
seems to be in a require form, offer module paths as candidates).