Add a fallback for re-provide-with-contract case
If no definition exists in the file reported by identifier-binding via
drracket/check-syntax, instead of returning #f a.k.a. "not found",
return the file:1:0. At least that is a helpful head start: If the
user can find the re-provide, and does another visit-definition on
_that_, it will probably take them to the true definition.

Again this is just a fallback. I'd still like to figure out a general
solution for making identifier-binding work in the face of contract
chaperone shenanigans, but I suspect that will need help from Robby
and/or Matthew.