Remove unnecessary use of markers
We do want to use markers for positions in the _values_ of the
properties racket-xp-def and racket-xp-use. That way they will adjust
automatically when the buffer edited.

But there is no reason to make markers to supply as the beg/end
position arguments when adding text properties. That is N/A. Text
properties are attached to the text and move with it, anyway.

For all I know `add-text-properties' is just calling `marker-position'
on them and otherwise ignoring them. In fact, do that ourselves when
supplying the def or uses markers as position arguments, just to
emphasize and defend against copypasta.

This doesn't fix a bug, per se. It definitely simplifies the code.