Also walk into submodules for completion candidates
Note: This still returns one set of symbols. It does NOT attempt to
return e.g. a tree where candidates are "scoped" by submodule paths,
such that the front end could limit candidates based on the innermost
submodule around point.

On the one hand, such a tree seems like it would be ideal. On the
other hand I'm not sure it's worth the effort? Furthermore, even if we
did so, I'm not sure how to do similar for the non-imported bindings
we get from drracket/check-syntax: Not even for module bindings and
the module path at point -- much less lexical bindings and "the
lexical scope at point"! At least this way, imported bindings are
handled consistently with other bindings: You get ALL the completion
candidates that might be valid ANYWHERE in the file. Is that too much?
Sometimes. Is that better than too little? Probably.