racket-tests/eventually: Add accept-process-output
This in an effort to figure out intermittent test failures on Travis
CI waiting for racket--repl-namespace-symbols to become non-nil.

However, running the tests locally has never failed for me yet, so,
I'm not sure how and why this would help. Ignoring that, my hypothesis
is, "Although sit-for waits for _input_, accept-process-output is what
might allow a command response to be processed".