Support multiple REPL buffers; closes #338
This was fairly short strokes from the work already done in the back
end related to REPL sessions now being TCP connections to the back end
server, and having it handle multiple such sessions. This just wasn't
exposed, yet, in the Emacs front end.

After mulling this over, and not wanting to write a whole bunch of
ceremonial code for "sessions", I realized that Emacs variable
semantics provide a simple, reliable way to handle this.

The variable `racket-repl-buffer-name' determines, in a `racket-mode'
edit buffer, the name of the `racket-repl-mode' buffer that it is
associated with. Like any Emacs variable, this may be global, or,
`setq-local' may be used to set a unique value for each buffer.

As a result, different strategies for associating one or more edit
buffers to one or more REPL buffers, all boil down to setting that
`racket-repl-buffer-name' variable as desired.

A new customization variable, `racket-repl-buffer-name-function', is
called when each `racket-mode' edit buffer is created.

- The default function is `racket-repl-buffer-name-shared', which sets
  the REPL buffer name to "*Racket REPL*". In other words, this is the
  status quo where all edit buffers share the same REPL.

- Another such function is `racket-repl-buffer-name-unique', which
  sets the REPL buffer name to something like "*Racket REPL:
  /path/to/foo.rkt*". In other words, somewhat like Dr Racket, each
  edit buffer gets its own, unique REPL buffer. (Currently there is
  no UX attempting to position them adjacently.)

- Another such function is `racket-repl-buffer-name-project', which
  names the REPL based on the projectile project name, if any. In
  other words, edit buffers in the same project share a REPL.

You get the idea. The user may supply any such function which sets the
desired REPL buffer name, possibly looking at `buffer-file-name' for
the `racket-mode' edit buffer as in the last two examples.