Handle errors with no source location
As a fallback, try to extract line:col information from exn-message.
Convert to the 1-based position used by the Emacs front-end.

Note that this is a last resort. But even when trying to do the right
thing using one of the various official ways, including using
exn:srclocs or exn:fail:syntax-exprs, there are still edge cases. For
example sometimes exn:fail:syntax-exprs will supply syntax -- but for
which syntax-position is #f.