dynamic-require => define-polyfill
We had one dynamic-require that is no longer needed because we're no
longer supporting Racket versions older than 6.0.

Change some others to use define-polyfill, when that is why they are
being used. Although partly the motivation is "use shiny new macro",
also this makes it easier to find polyfills and update or remove them
in the future.