racket-font-lock.el: implement and use racket-forward-sexp
This change substitutes the use of `forward-sexp' for
`racket-forward-sexp' which is sexp comment aware.

Currently `racket-forward-sexp' is only used for font locking, so I
put it in racket-font-lock.el. If there are use cases for it beyond
font locking then the helper functions can be moved.

Most of this implementation is directly from slime's handling of
reader-contitional forms. See
https://github.com/slime/slime/blob/cd8cc3c95c/slime.el#L7379-L7420
for reference.

Closes #432.