When editing text in an overlay, delete the overlay
Keeping the error affordance while you are editing to fix an error is
both distracting and discouraging ("you're still doing it wrong!").

Although maybe less important than it is for error annotations, this
is also good to do for bindings (defs, imports, uses, non-uses).
Whatever change the user is making, get out of their way until the
next annotation refresh.