Add file:1:0 fallback for plain find-definition, too
In the process, discovered bug trying to visit definition of
`module+`, because it was now reported as being at 1:0 of
racket/private/submodule.rkt instead of "not found". The bug is we
were matching on the outermost file module being the same as the file
name. Although that's true for #lang sources, it's not necessarily
true for files that directly consist of a module form. It turns out we
don't even care about the name of that outermost module, so the fix is
simply to disregard it i.e. match any name.

While doing so, took the opportunity to rename some things and
add/improve comments.