Improve command logging
Make a label from the command inputs, trimming any long strings (e.g.
the "code-str" argument to check-syntax is an entire Racket source
file).

Use that for log-racket-mode-info. Also use it to rename the command
thunk procedure, which in turn is how the thread gets its name, which
will appear when printing a thread value in logger output.