Revive the ability to cancel a check-syntax command
This got dropped back in commit e1f81f1. Although I don't remember
exactly why, and it may have been by accident, I think I also
convinced myself that it wasn't as valuable as I first imagined --
because the command was in fact fast enough, as well as the front-end
idle delay being good enough. Although that may be true, especially on
my relatively new laptop, I think it is straightforward to revive this
and it might help.