Require Racket 6.9 and define our own annotations-mixin
By defining our own annotations-mixin, we avoid needing to require the
build-trace% class from drracket/private/syncheck/traversals.

Also, build-trace% collects things in a vector, which we had been
de-duplicating and transforming. By using our own collector, we can
insert them directly into sets or hash-tables, in the first place, and
do less post-processing (mostly just changing into sorted lists for
the result to return to Emacs).

One gotcha: We want to override syncheck:add-unused-require when
available. That was added only c. Racket 6.9. There is no way (obvious
to me) to override something only if it exists in the super class, and
otherwise be a harmless no-op (as opposed to a runtime error). As a
result, as of this commit we now require Racket 6.9+. That is still a
relatively old version of Racket, and older than what is currently in
say Debian Stable. So I think that is acceptable (but if that poses a
hardship for some Racket Mode users, I could revisit this).