Fix def/drr followed by def
When refactoring this into an `or` expression I neglected to add a `t`
result to some sub-expressions. As a result, a `def/drr` was always
being followed by a `def`. Although in many cases it would "only" be
redundant and yield the same result -- in fact I only noticed this
while observing the racket-mode topic in racket-logger-mode -- it
might not always do so, and it makes things slower than necessary.