Add command to break; fixes #444
The switch to multiple REPLs meant that Emacs'
`comint-interrupt-subjob` won't break the user's REPL, because each
REPL is an Emacs network process; there is no "subjob" anymore.

Take the opportunity to split the `session` struct into `base-session`
struct -- which holds information about a session that is available
prior to module->namespace returning -- and a `session` struct that
derives from that, adding the namespace and repl-submit-predicate.
Update the `sessions` hash-table with a `base-session`, early, so that
the `thread` field is available to use with `break-thread` even during
a call to `module->namespace` (i.e. "while the user's program is still
running").