Use the :expected-result option for ert-deftest
That way we can say that a failing test is expected under certain
conditions.

Use :expected-result for:

- The macro expansion tests, which always fail on Racket 7.6.

- The racket-profile test, which fails rarely and intermittently on
  Travis CI. (I need to figure out why, eventually, but meanwhile
  allow the test to fail on CI, but require it to pass when tested
  locally.)

Keep in mind that ert-deftest is a macro and the :expected-result
value is evaluated at compile time. So, use the "satisfies" flavor to
evaluate a predicate at run time and after things like
`racket-program' are set appropriately.