Use command I/O for logger
Instead of running a separate little TCP server with its own port, use
command I/O: The logger level/topic configuration is now a command.
The logger output is now a "notification", similar to how we've been
handling debugger break notifications.

This leaves the back end using just one TCP port, for REPL sessions.
As of this commit, the front end specifies this port to the back end.
(In the next commit, we'll change that so that the back end uses an
ephemeral port, and tells the front end what port was chosen.)

See issue #452.