Recognize srcloc structs in REPL
We extend compilation-error-regexp-alist with a recognizer for srcloc
structs that get linkified and highlighted as INFO. This won't
distract you from normal work but will improve quality of life when
you work with parsers or readers that track source locations the
"Racket way". This tiny patch lets you immediately jump to
corresponding location.