Handle overlapping tooltips; closes #465
For mouse-over annotations use an interval-map where the values are
sets of strings.

As a result, mouse-overs like "Imported from typed/racket/base" and
"(-> Number Number)" can be combined correctly.

Using an interval-map matters because its interval-map-update*! should
correctly handle the case where items overlap partially.

Using a set of strings matters because syncheck:add-mouse-over-status
can be called with redundant items (which is why we were using a plain
mutable-set, originally).

In this approach, the front end racket-xp-mode should generally not
add help-echo property values except the "info" items provided by the
back end. So for example now the "Defined locally" tooltip is handled
by the back end. (This "unused require" help-echo is still added by
the front end but I think that's harmless and fine.)