Show explicit error message about `raco pkg install drracket`
See #468 and #471.

Although commit 0f1efa4 will at least surface an error message, for
this specific error we can provide a better message that explains what
people need to do.

This commit doesn't offer to run the command for the user, because it
can be called e.g. from an Emacs idle timer. Although that might be
fine, I'm not sure, and I'm not eager to introduce a possible new set
of bugs.