Comment out the racket-mode requirement
That way, racket-lexer-mode can be used e.g. with fundamental-mode for
.scrbl files.

It also might make sense to have .scrbl files handled by racket-mode.
But at the moment, my personal Emacs confif is still using the
scribble-mode package, and so I want to test .scrbl files by changing
them back to fundamental-mode and enabling racket-lexer-mode. In any
case, the Racket Mode _back end_ is required for racket-lexer-mode to
work, even if a buffer's major mode is not racket-mode.