Implement more sexp indent; connect to indent-line-function
At this point, things are mostly working correctly -- except when
concurrency of "update" and "indent" results in inconsistent state and
therefore wrong indent. Obviously that's a huge exception, and I'll
work on that next! But meanwhile it's still been useful to flesh out
the proof-of-concept of how to implement sexp indent using the
token-map interface.