Use set-port-next-location! for update re-tokenizing
I'd avoided this because we have no meaningful line and column
information to set. However, we don't actually use that information,
so it's fine to set it to something bogus like 1 and 0.

This is at least somewhat faster than doing read-byte-or-special from
the start of the input port to the desired start point.