WIP
- Huge optimization: Do not copy the entire original token
interval-map for comparison. Instead use a "copy on write" approach,
populating a temporary interval-map that stores the original intervals
only for those we actually have changed. This change makes things
fast even on the "torture test" drracket/private/unit.rkt which is
~5,000 lines.

- Supply an Emacs forward-sexp-function instead of marking open/close
tokens as open/close parens. This means we can support non-sexp langs.
However the utilization of forward-sexp-function in Emacs is not very
extensive, and things like paredit-mode do not work very well anymore.
So probably we'll need to let users decide whether to use this, or
not, on a per-buffer basis. If they choose not to, we can handle open
tokens like "#hasheq(" as "#hasheq" expression-prefix syntax and the
"(" as an open paren. Of course this only works when the close token
is e.g. a ")". But that is the case for most Racket sexp-based langs
AFAICT.

- As a result of the preceding item, we had to change backward-sexp
and forward-sexp to be able to return a "failed at position" value
instead of #f, to support how Emacs `up-list` catches a scan-error
from the forward-sexp function. A bit weird, but it works. This also
means that backward-sexp now must fail when it reaches an open paren,
rather than continuing indefinitely. That's OK, but it required
changing some of sexp-indent.rkt accordingly, and reviewing some
failing tests (some of which turned out to be bad tests).