Refactor documentation search
Build on the great work by Sergiu (now squashed down to the single,
previous commit) as well as insights from the resulting discussion.

Share more implementation and UI, both.

Among other things, ensure that when a search is done -- whether
because a specific binding could not be found, or because the user
wanted a search done in the first place -- that it is done the same
way, respecting the new racket-documentation-search-location
customization variable.

Rename some things and improve documentation.

Among other things, make it clear that the new command
`racket-documentation-search` is the go-to for a variety of situations
-- including where you want to assign some key binding to "always
search", as well as when `racket-xp-documentation` or
`racket-repl-documentation` are not available or won't work.

Because I wanted the command name to be `racket-documentation-search`,
I renamed the customization variable formerly of that name to be
`racket-documentation-search-location` (which is probably a clearer
name, anyway, than the name that I had originally suggested using).