Only rename identifiers that are textually equal
It is possible to have an arrow between two identifiers that are not
textually equivalent (definitely possible in DrRacket, and theoretically
possible in Racket Mode).

In the above event, renaming one identifier should not change the other.

See also racket/drracket#415 for a similar fix in DrRacket, and
racket/racket#3391 which would start to make Racket Mode behave
incorrectly. In particular, with the above PR, renaming
the identifier `a` in the following program should not rename
`all-defined-out` as well.

    #lang racket
    (provide (all-defined-out))
    (define a 1)