Automatically run certain submodules like DrRacket
The new customization variable `racket-submodules-to-run` says what
submodules to run. Its default value of `((test) (main)) is like the
default in DrRacket, and effectively closes #316.

See also #430, which was fixed by hardwiring an attempt to run the
main submodule if any. By contrast, this commit is more general and is
customizable, like DrRacket.