DRY issuing doc command and handling its response; fixes #494
Ever since the recent commit bec83a1, the "doc" back end command can
return false, which means no specific help could be found and the
front end should do a search. (The back end used to do a search in
this case automatically.)

Although the" helper function used by `racket-xp-documentation' and
`racket-repl-documentation' had been updated accordingly, the "lazy"
doc buttons created by `racket-xp-describe' and `racket-repl-describe'
for `racket-describe-mode' had not. (As a result, using such a button
could result in Emacs error that nil isn't stringp. Instead it should
open the Search Manuals page with the string for which a specific help
topic was not found.)

This commit fixes that oversight. Furthermore, instead of duplicating
this logic in various places, it is centralized in a new
`racket--doc-command' helper function.