Trim stack for REPL interactions
This commit extends the stack-trimming of commit 48741af to the REPL
thread, when it is waiting for the user to enter an interaction.

As a result, if they break with C-c C-c from racket-repl-mode this
will no longer print a stack trace into our interactions.rkt and
repl.rkt source (just like before commit 48741af changed how we do
stack-trimming).

Note: With Racket versions < 7.0 and our debugger REPL, we might see
an empty list after dropping the common tail in
continuation-mark-set->trimmed-context so handle that.