Make requires commands smart about modules when possible
racket-tidy-requires and racket-add-require-for-identifier now work in
terms of the innermost module around point. This could be a `#lang`
file's module (the only thing supported before) or another module
including a submodule nested to any depth.

This commit is not the minimal change necessary to support this
beause I took the opportunity to rename and refactor some things.

racket-trim-requires and racket-base-requires still only work on the
file module level -- and still warn when any module{*+} are detected
-- because the underlying macro-debugger/analysis/check-requires still
only support that. (Someday I'd like to submit a PR to enhance that...
assuming I am able to understand how it works, which seemed over my
head last I looked.)