Make identifier->bluebox a complete no-op on Racket 6.10
On 6.10 make-blueboxes-cache emits a very large number of log-warnings
about serialization failing due to "wrong shape". This in turn seems
to contribute to the problem of the REPL not starting quickly enough
or at all, and a test failure. At the moment I have such a situation
where all CI builds succeed except that for 6.10, and consistently.