More on issue #424: Add C-h for local "Search Manuals"
Intended as a convenience for when you have no idea which to pick and
need to read some documentation.

Note that this isn't necessarily exactly the same as doing M-x
racket-documentation-search, because that can be configured to use
docs.racket-lang.org, which might include modules that aren't
installed and can't be required. Instead this is more like doing a
"raco doc" at the command line.