Use the xref API
Define xref back ends for:

  - racket-mode
  - racket-xp-mode
  - racket-repl-mode

The old "-visit" and "racket-unvisit" commands are aliased/obsoleted
to the xref-find-definitions and xref-find-references front end
commands. Menus and keymaps and doc strings all updated.

Furthermore, the default xref-backend-references implementation greps
within the project. This is about as good as it gets, given Racket has
no project-wide database of references (check-syntax finds only
intra-file references). So we use that default back end generally, and
even as a fallback in the racket-xp-mode back end. Also, the
xref-find-references front end command provides a reasonably nice
major mode to view the references and navigate among them in
buffers (somewhat like next-error).