Add racket-trace-mode fed from vestige-trace logger topic
This is still just exploration.

This assumes a new Racket package called "vestige" which supplies
forms like racket/trace but which handles source locations and also
feeds a 'vestige-trace logger topic with events. That package isn't
even on GitHub as I type this, much less on pkgs.racket-lang.org.
Still experimental.

Note that such a package would not be a hard or even soft dependency
for Racket Mode per se, which simply looks for logger events in that
topic and expects them to be in a certain format; any code could
supply these logger events, not necessarily the "vestige" package.