racket-xp-mode: Show tail positions and targets
Handle expressions in tail position and their enclosing "target"
similarly to uses and definitions.

In the Racket back end, accumulate the check-syntax data to reveal the
1:many relation using a hash-table. Much like uses/defs. Also, handle
the fact that a single point might get mouse-overs saying it is both a
tail call and a tail target.

In the Emacs front end, use similar properties and point-motion
highlighting as we do for uses/defs.