let-bind print-{length level} around "%S" and pp; fixes #510
If the user or an obnoxious package `setq`s these they can limit our
use of print via "%S" in a format string. Indeed in issue #510
apparently a print-length of 10 was causing a "run" command
s-expression to be truncated -- as well as some lists in the M-x
bug-report output.

While we're updating racket-bug-report.el, also update the use of
backticks in the details section to instead use <pre> tags. At some
point GitHub changed this, and I've been manually editing bug reports
there to fix.