Update to use vestige "formals" and "values"
Also, remove the stuff highlighting parent levels in the source
buffers. It turns out to be distracting not useful. Also it's tricky
to get correct in all cases. The root cause for both? Juggling two
"dimensions" -- levels and caller vs. callee -- is too much cognitive
load as a user or as me the poor implementer.