Use new empty-span check-syntax annotations
Accommodate empty-span annotations. Although these could arise from
any syntax that set syntax-span to zero, that has been unusual until
recently. Starting circa Racket 7.9.0.14 check-syntax will produce
empty-span annotations for anything that expands to a use of `#%app`
or `#%datum`. Previously these were not annotated.

  https://github.com/racket/drracket/issues/445

A private class variable `use-empty-span-annotations?` controls this;
it could be exposed as e.g. a parameter if we want to make this user
configurable.

Setting this #f would make the situation similar to the previous
commit 6ec7c4c: Just ignore these annotations.

When #t, the default, we adjust their `end` to be `(add1 beg)`, i.e.
treat them as if they cover the first character. That seems to be the
least-worst way to handle them using things like data/interval-map and
Emacs text properties, for which empty intervals are either disallowed
or useless.

This commit also updates a couple spots in racket-xp.el where we had
an assumption that `(thing-at-point 'symbol)` is sufficient, which
isn't the case when the thing is e.g. a single char like ( or # or ".