Start to fold racket-trace-mode into racket-logger-mode
Idea is to eliminate racket-trace-mode; tracing is a special case of
logging.

This following recent updates to vestige where it provides some extra
information for logging generally -- such as depth -- not just
tracing. As a result, it makes sense for racket-logger-mode to be
enhanced to support the special case of tracing logging.