Various changes
- Add many more keybindings.

- Use (intern "<unknown>") as the symbol for unknown threads and
timings.

- Fix a missing 'invisible property.

- Add a 'racket-logger-message property, which covers the span that is
the logger message (as opposed to the other "fields" and prefixes).
Use this to in the racket-logger-{next previous}-item commands. More
reliable, especially in the face of multi-line messages. Also make the
commands skip invisible messages (when the buffer is filtered to show
items from just one thread).