Handle doing caller srcloc when there is no caller/trace srcloc
This can occur if a caller srcloc mark is in effect for a non-tracing
logging form.

For example, vestige's with-more-logging-info and/or
with-more-logging-depth forms  might supply such a mark. In that case,
even for a plain old logging form like log-debug, we could show source
where the form was used.