Indent Typed Racket result type on its own line; closes #516
This handles a result type on its own line, indenting it same as the
formals.

Although the original issue was about `lambda`, the same thing was
true of e.g. `define`, and this commit addresses both.

Note: Although this doesn't cause any existing indent tests to fail,
and although I added a couple new tests, I also didn't spend a ton of
time thinking through this. It's possible this doesn't work in all
cases, or worse, breaks something else. In that case I'll just revert
this commit. I think it's just a nice-to-have. Prior to this commit,
the behavior matches Dr Racket, which maybe isn't ideal but is
acceptable.