Add test for syntax cache memory use
This does a check-syntax for over 500 source files including many
Typed Racket sources, and observes memory use.

Checks that the net memory use is < 700 MB. That value simply comes
from what I'm observing, now. In other words this is a check that it
doesn't grow worse.

Also expose this and another slow test as a "test-slow" target in the
Makefile.