Syntax cache: Instead of a hash-table, use an association list
The list is ordered from MRU to LRU. The first `mru-to-keep` items are
non-evictable cache-entry structs; the remainder are evictable
cache-entry structs wrapped in ephemerons keyed by namespace.

The motivation here is to keep at least a small number of recent items
in the cache, even when they wouldn't be retained in an ephemeron by a
strong reference to a namespace. While still avoiding generally
unbounded growth in memory use as in issue #512.

[Keep in mind that ephemeron-value may return false quite quickly --
it is NOT the case that the value is retained until the next major GC.
A naive experiment where ephemeron-value is called immediately after
make-ephemeron may return non-false. But inserting, say, (sleep 1)
between the two will allow the ephemeron value to become false.]

Both cache-set! and cache-get traverse the entire list -- though just
once -- so they can:

1. Remove evicted items from the list.

2. Attempt to promote some to be non-evictable (generally impossible
   as described above, but harmless to try).

3. Definitely demote others to be evictable.