Simplify require-gui
There is no need to create an eventspace and set current-eventspace to
it. Just use a boolean flag.

Note: It is tempting to use racket/gui/dynamic here. Unfortunately
that would mean a package dependency on all of racket-gui-lib, which
would defeat most of the point of lazily loading the gui
on-demand. (The remaining, lesser part of the point is to start
faster, as well as to avoid any default GUI frame window and/or Racket
application icon that might be created on certain operating systems.)