Handle window-text-pixel-size returning zero in 28.0.50; fixes #544
Starting sometime after Emacs commit d10c96c, window-text-pixel-size
will return zero when given the exact start position of an overlay
with a 'display property.

It seems to work fine when given a start position that is the
beginning of the line. For our purposes here, it's fine to compare the
old vs. new widths of the entire line up to the end of the overlay, so
do that.