Add racket-{before after}-run-hook and racket-repl-clear
The ability to use racket-repl-clear with racket-before-run-hook
closes #149.

The hooks might turn out to be useful for other purposes.

Also add racket-repl-clear-leaving-last-prompt in case that's
something people would find useful as a command to run manually.

In order to avoid the cleared buffer starting with a blank line:

- Use fresh-line instead of newline, before re-running.

- Supply a new command to zero the column, so that fresh-line won't
think it's sitting at a non-zero column after a prompt and need to
emit a newline. Use this in racket-repl-clear.