Call zero-column! after reading current interaction
This should have been part of commit 72f2aa0, which stopped calling
zero-column! after displaying the prompt. That change was desirable so
that a "clear REPL every time" feature could work without a gratuitous
newline at the start.

However, calling zero-column! is necessary after reading an
interaction (user input at the REPL), so that what might next be
displayed will be correct. Although it might be N/A for a simple
print, it can matter for pretty-print, which uses port-next-location
to indent.