Eagerly load racket/gui/base when available
Try to dynamic-require racket/gui/base when the back end starts. If
that fails just ignore the error on the assumption that it is a
minimal Racket installation. Thereafter use racket/gui/dynamic to
handle both scenarios.

Note that racket/gui/dynamic is supplied by the `base` package ---
this does NOT create a dependency on the `gui-lib` package.

-----------------------------------------------------------------

In contrast the old approach was lazy: Use a module-name-resolver to
catch a user program's first attempt to require racket/gui/base,
require it ourselves, and restart the load of the user program.

One justification for the old approach was to avoid a hard dependency
on gui-lib --- but using dynamic-require and racket/gui/dynamic also
avoids this.

Another justification was to minimize startup time -- but any
difference seems barely measurable to me on semi-modern laptops.

A final justification was that merely requiring racket/gui/base could
cause OS window system artifacts to appear, such as a dummy window or
"Racket" application icon in a task bar or dock --- but I no longer
see this; maybe the behavior of the operating systems and/or Racket
has changed over the years. What I do see now:

Ubuntu 18.04 (Gnome) and Windows 10: No OS artifact appears until
`(send frame show #t)`. When all frames are hidden, the OS window
disappears.

macOS (Catalina): No OS artifact appears until `(send frame show #t)`.
When all frames are hidden, an "exec" icon in the application switcher
remains --- which isn't ideal but also isn't something Racket Mode
can/should handle.

-----------------------------------------------------------------

The background here, for any approach:

1. racket/gui/base cannot be instantiated more than once per Racket
process; attempting to do so causes an error.

2. A GUI program and its REPL need to run under an eventspace thread
via queue-callback, so that GUI events get a chance to be handled.