Added support for remote racket
When a racket session is created in a file visiting a remote server
via tramp the racket process is created in the remote server. If this
is the case the racket code running in the server is copied to the
remote server (if it is not there already). Said code is also modified
to accept connections from externa sources (security concern?).