Support multiple back ends; closes #553
This arises from PR #553, which adds the ability to use a Racket Mode
back end on a remote server: The use case is that remote .rkt files
are being edited using tramp, and we want to run the back end there as
well.

The original PR #553 uses the idea of copying racket--source-dir files
to the remote server, and, using `tramp-sh-handle-make-process'
instead of `make-process`, to start the back end process remotely.

This is an awesome PR but there are a couple issues:

1. The PR assumes a newer version of Emacs than our minimum:
`tramp-sh-handle-make-process' doesn't exist in Emacs 25.2.

2. Racket Mode currently assumes there is one, local back end. Various
other things need to be adapted. Furthermore, what if a user is
working with a mix of local and remote files? Pretty clearly, various
things won't work well unless/until we redesign to support multiple
back ends, each of which may be local or remote. This has the bonus of
letting people work with multiple /local/ back ends -- say, running
different versions of Racket itself.

This commit implements being able to use any number of local or remote
back ends simultaneously.